+++ src/main/java/org/cactoos/list/JoinedOf.java
@@ -23,33 +23,50 @@
  */
 package org.cactoos.list;
 
-import org.cactoos.CountOf;
-import org.cactoos.ScalarHasValue;
-import org.hamcrest.MatcherAssert;
-import org.junit.Test;
+import java.util.Iterator;
+import org.cactoos.func.StickyFunc;
 
 /**
- * Test case for {@link ConcatIterable}.
+ * A few Iterables joined together.
+ *
+ * <p>There is no thread-safety guarantee.
+ *
  * @author Yegor Bugayenko (yegor256@gmail.com)
  * @version $Id$
+ * @param <T> Type of item
  * @since 0.1
- * @checkstyle JavadocMethodCheck (500 lines)
  */
-public final class ConcatIterableTest {
+public final class JoinedOf<T> implements Iterable<T> {
 
-    @Test
-    public void transformsList() {
-        MatcherAssert.assertThat(
-            "Can't concatenate iterables together",
-            new CountOf(
-                new ConcatIterable<String>(
-                    new ArrayOf<>("hello", "world", "друг"),
-                    new ArrayOf<>("how", "are", "you"),
-                    new ArrayOf<>("what's", "up")
-                )
-            ),
-            // @checkstyle MagicNumber (1 line)
-            new ScalarHasValue<>(8)
+    /**
+     * Iterables.
+     */
+    private final Iterable<Iterable<T>> list;
+
+    /**
+     * Ctor.
+     * @param items Items to concatenate
+     */
+    @SafeVarargs
+    public JoinedOf(final Iterable<T>... items) {
+        this(new ListOf<>(items));
+    }
+
+    /**
+     * Ctor.
+     * @param items Items to concatenate
+     */
+    public JoinedOf(final Iterable<Iterable<T>> items) {
+        this.list = items;
+    }
+
+    @Override
+    public Iterator<T> iterator() {
+        return new ConcatIterator<>(
+            new MappedOf<>(
+                this.list,
+                new StickyFunc<>(Iterable::iterator)
+            )
         );
     }
